MCP/Wikipedia/Summarizer/requirements.txt
aiohttp>=3.8.0
fastmcp>=0.1.0
python-dotenv>=1.0.0
